Fix spanning tree arrayindexoutofbounds bug

+++ alpha/alpha-proc/src/main/java/org/neo4j/graphalgo/spanningtree/SpanningTreeBaseConfig.java
@@ -29,5 +29,5 @@ public interface SpanningTreeBaseConfig extends
     WriteConfig,
     RelationshipWeightConfig
 {
-    int startNodeId();
+    long startNodeId();
 }
